kernel: Fix k_sleep() in cooperative threads
Recent scheduler changes broke the case where k_sleep() (which was
unchanged and used direct timeout code) when called from a
unpreemptible thread, would still need to go to sleep.  Nothing was
setting an unrunnable bit in the state, the older scheduler was
detecting the case of the timeout being active.  The net effect was
that if the current thread could not preempt, k_sleep() would return
synchronously.

Rather than adding new logic or a new state bit, just implement sleep
in terms of the existing pend code, which allows for pending on a null
wait queue.

Patch also adds a useful assert used to catch this condition.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>