samples/hello_world: Revert SCHED_MULTIQ kconfig
This snuck in by accident in commit 9f06a3545044 ("kernel: Add the old
"multi queue" scheduler algorithm as an option").  Andrew Boie
actually caught it in code review, but I misread his criticism (I
thought he was asking about the other two tests that got the settings)
and handwaved it away.  This was just test code and not intended.
Revert.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>