kernel/timer: Spinlockify
This one has some logic changes.  Much of the timer API was expected
to be externally synchronized with the irq_lock().  Moving the
synchronization to an internal spinlock means that we can remove all
that external irq_lock() scaffolding.

But it also means we have to be careful about not assuming that the
lock is already held (which can race) or in calling newly-synchronized
functions with it already held (which will now deadlock -- spinlocks
aren't recursive like the irq_lock was!).

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>