arch/x86_64: cmake/qemu: Make "make run" work on x86_64
Like IAMCU, the x86_64 zephyr.elf reflects the native architecture and
not i386, which is a problem for qemu because it requires that the
loaded kernels be 386 protected mode binaries (i.e. they must match
the multiboot handoff ABI).

But unlike IAMCU, the existing qemu_machine_hack.py script is
unreliable with x86_64 binaries for reasons I haven't figured out.
Instead, this code does an objcopy followed by a relink of a proper
binary.  There's more cmake code and fiddlery involved, but the
mechanism should be more robust.

It also has the advantage of not munging the zephyr.elf file at "make
run" time, which is likely to be a desirable feature if we ever get
serious about repeatable builds.  The kernel passed to qemu in this
scheme is "zephyr-qemu.elf".

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>