arch/x86_64: Fix cmake include paths
The working directory for the C compiler is different between ninja
and make.  The include paths are fixed up, but that doesn't extend to
convincing the underlying assembler to search ".incbin" in the source
directory.  So put the stub files in to the generated/include
directory so they always get picked up.

This would almost allow me to drop the crazy "copy the C file" trick,
but I still can't figure out how to make a static source file depende
on "extra" stuff in cmake, so we still need an extra target mixed in.
The copy trick does that about as well as anything else.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>