arch/x86_64: Enter threads at _thread_entry() and not the user function
I made exactly the same mistake here that I did with xtensa.  The
_new_thread() function exposed by the architecture is given a
function, but it must not enter that function directly and instead
must provide it as an argument to the _thread_entry() wrapper.
Otherwise, when the user-provided function returns it pops off a
garbage returna address from the memory beyond the stack and explodes.

In theory the kernel could do that wrapping for us, but it doesn't.
Really it would be slicker here to avoid the extra function frame
entirely and instead push a single return address on the stack that
"returns" to a small assembly stub to invoke the cleanup code.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>