EXPERIMENT: Isolate timer API
The timer system is a mess, split in incestuous ways across multiple C
files and headers.

This rips it all out and replaced it with an empty/noop version in
time.c, with "minimized" headers to remove cruft.

Some things haven't been trimmed, as they're hard to find in many
cases (e.g. time ms_to_ticks family of functions, along with the
struct _timeout live in kernel.h).

But this is the broad structure of what the kernel and each driver is
expected to support.  Useful as a guide to writing a timer driver, or
for refactoring (or even rewriting) the subsystem.