tests/kernel/benchmarks/sys_kernel: Disable TICKLESS_KERNEL
This test was written to try to eliminate timer interrupts, but the
mechanism it chose (setting TICKS_PER_SECOND to 1) is dangerously
susceptible to overflow conditions on systems with fast cycle counters
and high timeout durations.

Actually the test passes just fine if you use a conventional tick rate
and use a tickless-capable driver (which eliminates interrupts too,
which is the whole point), but there's no way in a prj.conf file to do
an "if" to select that condition.  Just disable tickless to keep the
same behavior.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>