kernel: thread: correct type and member name of dlist node
Although sys_dnode_t and sys_dlist_t are aliases, their roles are
different and they appear in different positions in dlist API calls.
Change the name of the thread structure field used to link threads into
run queues and wait queues to reflect that it acts as a list node, not a
list head.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>