tests/kernel/schedule_api: Bump stack size and unify API
The new spinlock validation features combined with spinlockification
have increased stack usage a bit in CONFIG_ASSERT builds, but this is
a good feature we want to keep.  This test was bumping into limits, so
increase the size from 512 to 640 bytes.

Also unify the use of the STACK_SIZE macro, which was defined both in
a header and in files that included it (the compiler warns on
redefinition only if the values are different).

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>