kernel/spinlock: Add a depth count to spinlock validation
Right now the validation layer catches mismatched locking, but another
common gotcha is failing to release an outer nested lock before
entering a blocking primitive.  In these cases, the OS would swap
away, and the next thread (if any) to try to take the outer lock would
hit a "not my spinlock" error, which while correct isn't very
informative.

By keeping a single per-cpu nesting count we can check this at the
point of the swap, which is more helpful.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>