drivers/timer/cortex_m_systick: Fix MIN_DELAY computation
The minimum delay was being set when the ticks argument was zero, but
not being checked on the final delay computed.  This worked as long as
interrupts are always delivered promptly, but when very delayed the
wraparound logic an produce arbitrarily silly delay values, and it was
possible to fool the driver into setting an interrupt sooner than the
limit.

Also, the loop in elapsed() was needless.  The intent was that you
check both the CTRL register and VAL register in a loop and test the
VAL read against the new one to make sure the counter didn't roll
over, then do it again.  But the only thing we were reading the CTRL
register for WAS the rollover bit!  And it clears on read anyway, so
we were only computing a bitwise OR of both reads anyway.  Basically,
the hardware already does what we wanted.  Take out the third register
read, it was useless.  But it is important to read CTRL after VAL so
it properly reflects the rollover state.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>