arch/arm: Fix race in syscall entry
The exit from the SVC handler into _arm_do_syscall() was racy.  The
exception return inherently unmasks interrupts, meaning that any
interrupt queued while the SVC handler was running happens BEFORE
_arm_do_syscall() starts.  If that interrupt happened to cause a
scheduling event, our current context can be swapped away and the
_current pointer that we find in memory may not match the one of the
thread that invoked us!

Rather than try to find a way to keep interrupts masked across that
boundary, just put _current into a register when entering the thread
mode handler.  This forces an extra spill of the register in the
usermode syscall wrapper, but actually saves four instructions in the
handler, so it's even a performance win.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>