drivers/timer: Work around qemu "time warp" behavior
Qemu reflects a host clock rate inside its guests, which actually
helps us get coverage for timing-related bugs that would be invisible
in a deterministic system.  But the downside is that emulator is
running as a regular process in a host OS, and that host OS, when
loaded, can decide not to run the guest for large chunks of time
(i.e. a host OS scheduling quantum, unsurprisingly about the same size
as our zephyr ticks).  When this happens, the guest appears to "time
warp" into the future.  Timer interrupts arrive very late, tests get
confused as to why things didn't happen when they expected, and we get
spurious failures when running tests in a loaded environment like
sanitycheck and CI.

So this modifies the hpet, riscv and xtensa drivers to detect this
condition in the ISR, and skew the internal clock to reflect the time
the late interrupt "should" have arrived.

This works as long as qemu is internally consistent with its
timekeeping.  On older versions, I (think I!) have seen qemu get
things really wrong and actually run interrupt-enabled guest code in a
context where the interrupt should have arrived long ago and hasn't.
But I can't reproduce these issues with the qemu (3.0.50) in the
Zephyr 0.9.5 SDK.

That feature replaces the previous CONFIG_QEMU_TICKLESS_WORKAROUND
kconfig, so that has been removed.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>