tests/kernel/interrupt: Correct timing usage
This tests hard codes a 100 Hz tick rate, but it sets a k_timer for a
DURATION value of 5ms, which cannot be done with a precision, then
checks it against an unrelated k_busy_wait() time of exactly 10ms,
which is right at the boundary of what is possible.  Playing with
timer stuff can push this over the edge easily.

Correct DURATION to reflect "one tick", and spin for that value, plus
exactly one tick of slop before testing to see whether the timer ISR
ran.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>