tests/kernel/fatal: Fix clobbered k_timer
The tests all share a stack, so by putting the k_timer on the stack
we're potentially allowing it to be clobbered by a following test
before it expires.  Make it static.

Oddly this "worked" when written, becuase it initialized the timer
after the stack clobber, even though the rest of the stack was already
garbage, and because the timing just happened to expire before the
next test got to it.  Mucking around with timeout expirations exposed
the bug.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>