kernel/sched: Fix race vs. queued _current in SMP
In SMP, a CPU's _current thread does not live in the queue, making it
invisible to other CPUs by design.  But there is an unlockable time
period between enqueing the current thread (e.g. as part of a swap,
though there are multiple places where this can happen) and the
arch_switch() call that actually saves the thread state.  We can lock
the current CPU via interrupt state (which gets atomically set as part
of the context switch), but we can't prevent another CPU from seeing
and attempting to run a thread in this intermediate state.

But we can exploit the existing arch_swtich() API which does a write
to the output switch handle.  On SMP, a thread that has added _current
to the queue will put a NULL into its handle field, and the scheduler
will spin on that to be written before using such a thread as a target
via interrupt return or swap.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>