serial: ns16550: device config struct be non-const when PCIE=y
When PCIe is enabled for UART, the port address is probed during
initialization and is written back into the struct. When MMU is
enabled, the struct is located in a read-only region, and any
write will cause a fault. So work around the issue by not
marking the struct as const when PCIE is enabled.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>